split WebinosService into its own file
changes concatenation order when building webinos.js, the discovery code
now is at the very end. since the service modules need the WebinosService
interface, this code is split into its own file so it's availabe to them.

the call to instantiate the discovery is now also in webinos.servicedisco.js.

Jira-issue: WP-848