Allow certificate configurations to reference external files
This commit adds a function to the configuration.js module
to allow a pzh configuration to point to a file containing
a PEM certificate rather than forcing the PEM certificate
to be inside the JSON structure.  This will make it easier
to deploy the PZH Web Server without re-editing the
configuration files each time.

Jira issue: WP-965